requirements.txt
apache-airflow==2.0.2
kaggle
dvc
google-cloud-storage
requests
dvc[gs]